Fix trying to delete supplemental folder if it doesn't exist.

+++ Results/utils.py
@@ -83,4 +83,5 @@ def delete_all_outputs():
     for model in [DailyControls, DailyByZone, DailyByProductionType, DailyByZoneAndProductionType, UnitStats, ResultsVersion]:
         model.objects.all().delete()
     SmSession.objects.all().update(iteration_text = '', simulation_has_started=False)  # This is also reset from open_scenario
-    shutil.rmtree(workspace_path(scenario_filename() + "/" + "Supplemental Output Files"))
+    if os.path.isdir(workspace_path(scenario_filename() + "/" + "Supplemental Output Files")):
+        shutil.rmtree(workspace_path(scenario_filename() + "/" + "Supplemental Output Files"))
